[model  server] Test `defaultDeploymentMode` configuration in DSC and configmap (#133)
* Create size-labeler.yml

* Delete .github/workflows/size-labeler.yml

* model mesh - add auth tests

* xx

* add dsc default deployment tests

* do not run sessionfinish on collectonly

* add dsc

* update dsc with funtion

* set deployment_mode optional

* Address comments

* fix storage config

+++ conftest.py
@@ -207,6 +207,9 @@ def pytest_report_teststatus(report: CollectReport, config: Config) -> None:
 
 
 def pytest_sessionfinish(session: Session, exitstatus: int) -> None:
+    if session.config.option.setupplan or session.config.option.collectonly:
+        return
+
     base_dir = py_config["tmp_base_dir"]
     LOGGER.info(f"Deleting pytest base dir {base_dir}")
     shutil.rmtree(path=base_dir, ignore_errors=True)
